Reduce huffman decode table size

+++ src/hpack/huffman/mod.rs
@@ -7,7 +7,7 @@ use bytes::{BufMut, BytesMut};
 
 // Constructed in the generated `table.rs` file
 struct Decoder {
-    state: usize,
+    state: u8,
     maybe_eos: bool,
 }
 
@@ -76,7 +76,7 @@ impl Decoder {
     // Decodes 4 bits
     fn decode4(&mut self, input: u8) -> Result<Option<u8>, DecoderError> {
         // (next-state, byte, flags)
-        let (next, byte, flags) = DECODE_TABLE[self.state][input as usize];
+        let (next, byte, flags) = DECODE_TABLE[self.state as usize][input as usize];
 
         if flags & ERROR == ERROR {
             // Data followed the EOS marker
